pandoc & metadata
