+++ src/command/render/layout.ts
@@ -9,7 +9,7 @@ import { Document, Element } from "deno_dom/deno-dom-wasm.ts";
 
 import { kPageWidth } from "../../config/constants.ts";
 import { Format } from "../../config/format.ts";
-import { Metadata } from "../../config/metadata.ts";
+import { Metadata } from "../../config/types.ts";
 import { resourcePath } from "../../core/resources.ts";
 
 export function layoutFilter() {
