+++ src/command/render/pandoc.ts
@@ -36,11 +36,7 @@ import {
   kTextHighlightingMode,
   SassBundle,
 } from "../../config/format.ts";
-import {
-  isQuartoMetadata,
-  Metadata,
-  metadataGetDeep,
-} from "../../config/metadata.ts";
+import { isQuartoMetadata, metadataGetDeep } from "../../config/metadata.ts";
 import { binaryPath, resourcePath } from "../../core/resources.ts";
 import { pandocAutoIdentifier } from "../../core/pandoc/pandoc-id.ts";
 import {
@@ -52,7 +48,7 @@ import { ProjectContext } from "../../project/project-shared.ts";
 import { deleteProjectMetadata } from "../../project/project-context.ts";
 import { deleteCrossrefMetadata } from "../../project/project-crossrefs.ts";
 
-import { removePandocArgs, RenderFlags } from "./flags.ts";
+import { removePandocArgs } from "./flags.ts";
 import {
   generateDefaults,
   pandocDefaultsMessage,
@@ -81,7 +77,7 @@ import {
 import { sessionTempFile } from "../../core/temp.ts";
 import { cssImports, cssResources } from "../../core/css.ts";
 
-import { RunPandocResult } from "./types.ts";
+import { PandocOptions, RunPandocResult } from "./types.ts";
 import { compileSass } from "./sass.ts";
 import { crossrefFilterActive } from "./crossref.ts";
 import { kQuartoHtmlDependency } from "../../format/html/format-html.ts";
@@ -91,41 +87,10 @@ import {
   formatHasCodeTools,
   keepSourceBlock,
 } from "./codetools.ts";
+import { pandocMetadataPath } from "./render-shared.ts";
+import { Metadata } from "../../config/types.ts";
 
 export const kMarkdownBlockSeparator = "\n\n<!-- -->\n\n";
-
-// options required to run pandoc
-export interface PandocOptions {
-  // markdown input
-  markdown: string;
-
-  // original source file
-  source: string;
-
-  // original metadata
-  metadata: Metadata;
-
-  // output file that will be written
-  output: string;
-
-  // lib dir for converstion
-  libDir: string;
-
-  // target format
-  format: Format;
-  // command line args for pandoc
-  args: string[];
-
-  // optoinal project context
-  project?: ProjectContext;
-
-  // command line flags (e.g. could be used
-  // to specify e.g. quiet or pdf engine)
-  flags?: RenderFlags;
-
-  // optional offset from file to project dir
-  offset?: string;
-}
 
 export async function runPandoc(
   options: PandocOptions,
@@ -450,10 +415,6 @@ export async function runPandoc(
   } else {
     return null;
   }
-}
-
-export function pandocMetadataPath(path: string) {
-  return pathWithForwardSlashes(path);
 }
 
 async function resolveExtras(
