+++ src/command/serve/serve.ts
@@ -33,7 +33,7 @@ import {
   projectIgnoreRegexes,
 } from "../../project/project-context.ts";
 import { inputFileForOutputFile } from "../../project/project-index.ts";
-import { kProject404File } from "../../project/project-resources.ts";
+import { kProject404File } from "../../project/project-shared.ts";
 
 import { websitePath } from "../../project/types/website/website-config.ts";
