+++ src/config/types.ts
@@ -1,0 +1,10 @@
+/*
+* types.ts
+*
+* Copyright (C) 2020 by RStudio, PBC
+*
+*/
+
+export type Metadata = {
+  [key: string]: unknown;
+};
