+++ src/project/project-resources.ts
@@ -14,13 +14,12 @@ import { resolvePathGlobs } from "../core/path.ts";
 import { kCssImportRegex, kCssUrlRegex } from "../core/css.ts";
 
 import {
+  kProject404File,
   kProjectOutputDir,
   kProjectResources,
   ProjectConfig,
 } from "./project-shared.ts";
 import { kQuartoIgnore } from "./project-gitignore.ts";
-
-export const kProject404File = "404.html";
 
 export function projectResourceFiles(
   dir: string,
