+++ src/project/types/website/website-analytics.ts
@@ -7,7 +7,7 @@
 import { Document } from "deno_dom/deno-dom-wasm.ts";
 import { join } from "path/mod.ts";
 import { kTitle } from "../../../config/constants.ts";
-import { Metadata } from "../../../config/metadata.ts";
+import { Metadata } from "../../../config/types.ts";
 import { projectTypeResourcePath } from "../../../core/resources.ts";
 import { sessionTempFile } from "../../../core/temp.ts";
 import { ProjectContext } from "../../project-shared.ts";
