+++ src/project/types/website/website.ts
@@ -13,9 +13,12 @@ import { resourcePath } from "../../../core/resources.ts";
 import { dirAndStem } from "../../../core/path.ts";
 import { isHtmlContent } from "../../../core/mime.ts";
 
-import { ProjectContext } from "../../project-shared.ts";
-import { projectOffset, projectOutputDir } from "../../project-shared.ts";
-import { kProject404File } from "../../project-resources.ts";
+import {
+  kProject404File,
+  ProjectContext,
+  projectOffset,
+  projectOutputDir,
+} from "../../project-shared.ts";
 
 import {
   ProjectCreate,
